pyproject.toml
[tool.poetry]
name = "graphql-dynamodb"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"
packages = [{include = "lambdas"}]

[tool.poetry.dependencies]
python = "~3.9"
strawberry-graphql = {extras = ["fastapi"], version = "^0.155.0"}
mangum = "^0.17.0"
pynamodb = "^5.3.4"


[tool.poetry.group.deployment.dependencies]
aws-cdk-lib = "^2.60.0"
constructs = "^10.1.222"


[tool.poetry.group.dev.dependencies]
black = "23.1.0"
pre-commit = "^2.21.0"
boto3 = "^1.26.50"
boto3-stubs = {extras = ["lambda"], version = "^1.26.55"}
faker = "^16.6.0"
ruff = "0.0.231"
strawberry-graphql = {extras = ["debug-server"], version = "^0.155.0"}


[tool.black]
target-version = ["py39"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
